Update and rename 2447.cpp to 2447_2.cpp
